stacks/train_model/requirements.txt
scikit-learn==1.5.2
google-cloud-storage==2.18.2
db-dtypes==1.3.0 
pandas==2.2.3 
google-cloud-aiplatform==1.69.0
PyYAML==6.0.2
python-box==6.0.2
types-PyYAML==6.0.12.12